solved? who knows...

+++ basic_threads/primes/task.h
@@ -57,7 +57,7 @@ public:
         auto t2 = std::chrono::high_resolution_clock::now();
         auto t3 = std::chrono::high_resolution_clock::now();
         time_waiting_for_mutex_ = time_waiting_for_mutex_ + t2 - t1;
-        for (int i = 0; i < R - L + 1; ++i)
+        for (uint64_t i = 0; i < R - L + 1; ++i)
             if (isPrime[i])
                 primes_.insert(i + L);
         auto t4 = std::chrono::high_resolution_clock::now();
